requirements.txt
# Core dependencies
boto3>=1.34.0
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Social media APIs
pytumblr>=0.1.2
atproto>=0.0.46

# AI caption generation via AWS Bedrock (uses boto3)
# No additional dependencies needed - uses existing AWS credentials
